requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dnspython==1.12.0
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
matrix-client==0.0.6
pycrypto==2.6.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
slackclient==1.0.9
-e git+https://github.com/lampholder/slack_sync.git@df2b88c82a3a901dde5ddf149dc69dc47f3dbeea#egg=slacksync
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
